feat(front): prettier header (#280)
* update footer -> put former header links

* refactoring navigation bar: fix dropdown menu when loggedin + update navbar style + add icons

* move header footer to components

* header i18n

* update package lock

* e2e: test fix navigation

* e2e: renable tests

* footer: remove header links

* update package lock

* Revert "update footer -> put former header links"

This reverts commit 3d7b0b27c575659275357537ab637f27946d1bfe.

* e2e tests: cypress remove force

* fix responsive manifesto page

* update package lock
