+++ components/layout/ScreenContainer.tsx
@@ -31,7 +31,7 @@ export const ScreenContainerCentered: React.FC<ScreenContainerProps> = ({
   isSignedOutModal,
 }) => {
   return (
-    <div className="h-screen flex flex-col items-center justify-center">
+    <div className="flex grow flex-col items-center justify-center">
       <main className="flex flex-col w-full justify-center items-center mb-10 sm:mb-0">
         <div className="mt-8 mx-5">{children}</div>
       </main>
